[cpu] add simulation of the TX-2's Exchange Element.
TX-2 instructions which interact with (main, i.e. STUV) memory do so
via the mediation of the Exchange Element.  The Exchange Element
permutes the bits of the data as it is transferred into or out of
memory.  It can also suppress transfer of part of the word, and
perform sign extension.  Whether and how this happens is configured
via the "configuration" field of the current instruction.

The mapping between an instruction's configuration and the actual
actions of the exchange unit is determined by what the F-memory holds
at the location identified by the instructions configuration field.
There is a standard setup for this (decribed in table 7-2 of the User
Handbook), but it is modifiable by the programmer in software.