[cpu] add initial implementation of the TX-2's control unit.
This is very basic; it only implements the instructions SKX, DPX, JMP,
JPX, JNX, SKM and SPG, the X-memory, F-memory and the
CODABO (i.e. boot) feature.

This is enough, however, to execute the initial boot code which clears
memory and sets up the F-memory.  There is no support yet for I/O, so
we can't execute last the part of the boot code which loads a binary
from paper tape and runs it.