Exchange unit: is directional, fix sign extension.
Make the exchange unit behave differently in the E->M and M->E
directions.  This follows the explanation in chapter 13 of the Users
Handbook.

Also fix sign extension.  It doesn't happen in the E->M direction and
the sign bit of a quarter is 0o400, not 0o300.

Add test cases for the RSX instruction which correspond to the
examples in the instruction definition in the Users Handbook.