Implement registers A, B, C, D and opcodes LDA, LDB, LDC, LDD, LDE.
The metabit of registers A, B, C, D is not correctly implemented.
Thes registers actually don't have one.  In the TX-2, manipulation of
the metabit of these registers actually affects only the M register.

This may well be probably program-detectable, depending on whether
loading an AE register into the exchange unit (i.e. the M register)
resets the meta bit or leaves it set to the existing value.  If the
latter, setging the meta bit of the A register will apparently also
set the meta bit of the B register.

Perhaps also for the modes in which the system sets the metabit of
words which were used as instruction operands.