Initial implementation of an alarm panel.
This has some quality-of-implementation issues:

- coupling between the simulated machine and UI components is via
  globals in machine.ts, and there should probably be a class
  responsible for this mapping (e.g. an MVC controller).
- there are too many similar-looking alarm-status interfaces, because
  there is no clean separation between machine.ts and AlarmPanel.
- Handling of props and status in AlarmPanel may not correspond
  with React best-practice.

This code also reveals an issue in the simulator: if OCSAL is masked,
the simulator attempts to continue execution with an unchanged value
of P, which of course will never proceed (since OCSAL simply results
again).  This likely does not correspond with the behaviour of the
original machine.