Use controller (as in MVC) objects.
This is still quite unsatisfactory.  We're using callbacks for the
model to update the UI (e.g. the "Run" checkbox) and also using
callbacks for the UI to update the model (i.e. for the "Run" checkbox
to have an effect).  This works, but is ugly, there must be a better
way.