Style the status panels with external CSS.
We run the build twice if necessary (due to a missing CSS type file).

The src/styles.scss.d.ts file is both generated by the compilation
process and an input to the compilation process.  Running the build
twice ensures that when the src/styles.scss.d.ts file is missing, the
first build fails but generates it, and so the second build can
succeed because the file is now in place.