Add a second canvas for hit-detection.
The principle behid this is that the items on this second canvas have
the same positions and dimensions as the keys drawn on the first, but
each has a unique colour.  When there is a hit at (x,y) on the first
canvas, we look up the colour of the pixel at (x,y) in the second
canvas, and its distinctive colour allows us to deduce which key is
at (x,y) in the first canvas.

Key-press events are not yet handled, and this second canvas is
currently visible (for easier debugging) which would not be the case
in real use.