Update github actions to v3 (Node v12 is deprecated in actions).
We update (in this change) only actions maintained by github.

The update is necessary to migrate off Node v12, because this is going
out of support per the announcement at
https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12.

Some other actions we use still need an update so we expexct this
change to reduce but not eliminate the warning messages.