Perform CI steps on all components in the workspace.
The purpose of this change is actually to (re-) test the configuration
change made in 86702af1790d0474e4437d952346a694ff880722, which wasn't
tested at the time I pushed that change because dtolnay/rust-toolchain
was not, at that time, in the list of permitted workflows.  I can't
see a way to re-test this other than pushing another change.