[cpu] Improve invariant in VMemory::read_access().
Previously read_access() returned Result<Option<&mut MemoryWord>,
MemoryOpFailure>.  This meant in theory that it could return Ok(None),
signalling that a write access had been attempted to a read-only
location. Since that doesn't make sense for a read, we change it to
return Result<&mut MemoryWord, MemoryOpFailure>, eliminating the
possibility of incorrect semantics.  Because access() needs to have a
single return type, we map the result of the read operation so that it
always returns a non-empty Option.