Fix mapping of LW code 0o20. Now n, was ≈, Almost Equal To (U+2248).
This symbol was simply hard to read in the character set table in the
Users Handbook.  Because there is not a complete set of lower-case
letters on the Lincoln Writer, the fact that 'n' was missing wasn't
obviously wrong.  The documentation is quite subject to
misinterpretation:

- Users Handbook: blob that could be ≈ or n.
- Progress report diagram: symbol that could be η (eta) or n.

The "n" interpretation is the only one which is consistent with both
documents.

This change also introduces a test which checks that the character-set
translation in the base crate (charset.rs) is consistent with the
labels on the keyboard in the web UI (in keyboard.rs).  This test
catches this bug.

Fixes: #111