[assembler] Initial nnew lexer based on recognising glyphs.
This new lexer first converts the input character stream into a stream
of glyphs enum values.  This means that, for example "@sup_eps@" and
"ᵋ" will be handled identically (except that they have different
spans).

The idea is to keep the existing Logos-based low-level lexer (which
recognises comments for example) but replace the existing (also
Logos-based) high-level lexer (which recognises numeric literals for
example) with this glyph-based one.  However, this change doesn't make
that swap.