[assembler] minor bugfixes to glyph parser.
The main change is that some character codes (those for aquare,
circle, underbar and overbar) do not advance the carriage.  Therefore
we don't use ASCII '_' to yield the underline/underbar glyph.

Also a bugfix to test cases; @plus@ should actually be @add@,
following Jurij's convention in sub.py.

I believe most of the remaining test failures are due to
inconsistencies of span.  This is probably due to a problem with how
the glphy-based parser measures spans.  It sometimes uses character
counts when in reality it should be counting bytes (to be consistent
with how &str slices work).  Hence it probably shouldn't be using
constructs like s.chars().enumerate().