[assembler] We can't correctly implement tabs, so reject them.
The issue is that spaces are allowed in symexes but tabs are not.  To
correctly handle this in the parser, we'd need to update the grammar
to allow spaces almost everywhere, except between two symex syllables.
This would make the grammar so much more complex that it seems simpler
to disallow the use of tab, and require spaces instead.

When we see a tab, we issue an error message explaining what to use
instead, for the most obvious possible user intents.