[assembler] Implement the "pipe construct" (ADXₚ|ₜQ).
This is described in the User Handbook section 2-2.8 "SPECIAL
SYMBOLS": "ADXₚ|ₜQ" should be equivalent to "ADXₚ{Qₜ}*".

Note that the pipe symbol should actually be in subscript; this is
correctly implemented in the code, but since Unicode has no subscript
pipe symbol, we can't render it correctly in a git comment or in
OPEN-QUESTIONS.md.