[assembler] Comma is a macro terminator, dot is not.
They are actually hard to distinguish in the copy of the Users
Handbook I have.  But, looking at page 011 of Leonard Kleinrock's
listing for his network simulator, the `HP OS` macro is definitely
using as a separator a symbol that lives on the line.  While if you
look a little further below on the same page, the third instruction in
the body of the `MV MX` macro body contains both a dot and a comma.
The dot is definitely above the line and looks rounder.  So I conclude
that the separator character is a comma.