[assembler] Tags defined in RC-words are not local to the RC-word.
Tags defined in RC-words may not be used for M4's editing instuctions,
but nevertheless they are not locally-scoped.  This is demonstrated by
the example in section 6-4.7 of the User's Handbook, which looks like
this:

```
☛☛DEF TBS|α
 α
 α
 α
 α
 α
☛☛EMD
100|
USE->     LDA {TS->TBS|0}  ** 5 BLANK RC WORDS
          LDA TOMM
          STA TS+3
```

You will see above that the definition of the tag `TS` is inside an
RC-word, but _not_ inside a macro body.

The example explains that the above code snippet expands to:

```
100|
USE ->    LDA {TS-> TBS| 0}              |002400 000103|000100
          LDA TOMM                       |002400 000110|   101
          STA TS+3                       |003400 000106|   102
TS ->     TBS 0
          0                              |000000 000000|   103
          0                              |000000 000000|   104
          0                              |000000 000000|   105
          0                              |000000 000000|   106
          0                              |000000 000000|   107
TOMM ->   0                              |000000 000000|000110
```