[assembler] lexer: use Option<Token> instead of Option<Result<Token>>
The Token enum already has an Token::Error that we use to signal an
error, so the Err variant of the result was only used to complain
about finding a tab character.  Instead, update Token::Error to be
able to signal that and eliminate the use of Result.  Then we have
just one way to signal that an error has occurred.