[cpu] Raise ROUNDTUITAL alarm if the AOP instruction is used.
Also OCSAL when the sub-code is of opcode 04 indicates that an
undefined instruction is being executed.

Opcode 04 (sometimes called OPR) apparently encodes both the IOS
instruction and the AOP instruction, according to bits 2.7 and 2.8 of
the instruction word.  If bit 8 is set, the instruction is invalid, so
we raise OCSAL.  If bit 7 is unset, bit 8 distinguishes IOS (0) and
AOP (1).  Since AOP isn't implemented, it causes ROUNDTUITAL to be
raised.