Add a workflow for deploying Rust documentation to github pages.
We're not planning to publish these crates to https://crates.io,
primarily because we're not targeting library use-cases (and, for
example, we don't want to manage API compatibility issues).

Nevertheless having Rustdoc documentation will likely help people
(for example, me) understand the code.

Becasue the audience for these docs is the developers of the code, we
also document private items.